--- database_api/logs/06-06-2024.log
@@ -1,5 +1,0 @@
-[06-06-2024 17:50:47]:💬 Информация -> Запрос от 127.0.0.1:49204
-[06-06-2024 17:56:24]:💬 Информация -> Запрос от 127.0.0.1:55350
-[06-06-2024 17:56:24]:💬 Информация -> адрес запроса /
-[06-06-2024 17:56:24]:💬 Информация -> адрес запроса /favicon.ico
-[06-06-2024 17:56:42]:💬 Информация -> адрес запроса /register_service
